refactor(web): reorganize data dir
