+++ service/vspo-schedule/web/src/constants/navigation.ts
