+++ service/vspo-schedule/web/src/constants/timeframes.ts
@@ -1,0 +1,8 @@
+import { Timeframe } from "@/types/timeframe";
+
+export const timeframes: { label: string; value: Timeframe }[] = [
+  { label: "1日", value: "1day" },
+  { label: "1週間", value: "1week" },
+  { label: "1ヶ月", value: "1month" },
+  // { label: "3ヶ月", value: "all" },
+];
