+++ service/vspo-schedule/web/src/data/freechat-video-ids.ts
@@ -1,0 +1,17 @@
+export const freeChatVideoIds = [
+  "tFFq_vUblrs",
+  "Rfuu2gkj18w",
+  "bx1-cTN0Zas",
+  "HC8rMmTWBGk",
+  "lSvLABLEhO0",
+  "CIRvsQ7XMoM",
+  "RasCgfRnTOc",
+  "l4_7qk5VE50",
+  "7-rmkxy7SSg",
+  "uZduRVQgZgo",
+  "JkuHbeoqu64",
+  "Fs4QLxzm-Fc",
+  "pT6aeI5S1Kk",
+  "sKJ5moF3X8Y",
+  "7-rmkxy7SSg",
+];
