+++ service/vspo-schedule/web/src/data/mocks/clips.ts
