+++ service/vspo-schedule/web/src/data/mocks/events.ts
