+++ service/vspo-schedule/web/src/data/mocks/freechats.ts
