+++ service/vspo-schedule/web/src/pages/twitch-clips.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from "react";
 import { GetStaticProps } from "next";
 import { Clip } from "@/types/streaming";
 import { filterByTimeframe, formatWithTimeZone } from "@/lib/utils";
-import { mockTwitchClips } from "@/data/clips";
+import { mockTwitchClips } from "@/data/mocks/clips";
 import { Box } from "@mui/system";
 import { NextPageWithLayout } from "./_app";
 import { Loading, SearchDialog } from "@/components/Elements";
