+++ service/vspo-schedule/web/src/types/timeframe.ts
@@ -1,0 +1,1 @@
+export type Timeframe = "1day" | "1week" | "1month";
